+++ packages/ui-contexts/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rocket.chat/ui-contexts
 
+## 12.0.0-rc.4
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.0.0-rc.4
+  - @rocket.chat/rest-typings@7.0.0-rc.4
+  - @rocket.chat/ddp-client@0.3.9-rc.4
+  </details>
+
+## 12.0.0-rc.3
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.0.0-rc.3
+  - @rocket.chat/rest-typings@7.0.0-rc.3
+  - @rocket.chat/ddp-client@0.3.9-rc.3
+  </details>
+
 ## 12.0.0-rc.2
 
 ### Patch Changes